Create term controller; Connect route to app; Add module existence ckeck to term service;

+++ src/services/term.service.ts
@@ -6,32 +6,47 @@ import { TermDto } from '../dto/term.dto';
 import { ITermService } from '../interfaces/services/term.service.interface';
 import { ITermRepository } from '../interfaces/repositories/term.repository.interface';
 import { Term } from '../models/term.entity';
+import { IModuleRepository } from '../interfaces/repositories/module.repository.interface';
 
 @injectable()
 export class TermService implements ITermService {
 	constructor(
 		@inject(TYPES.ITermRepository) private termRepository: ITermRepository,
+		@inject(TYPES.IModuleRepository) private moduleRepository: IModuleRepository,
 		@inject(TYPES.IConfigService) private configService: IConfigService,
 	) {}
 
 	async createTerm(
+		userId: string,
 		moduleId: string,
 		{ term, definition, status, isStarred }: TermDto,
 	): Promise<TermModel | null> {
+		const existedModule = await this.moduleRepository.getModuleById(userId, moduleId);
+		if (!existedModule) {
+			throw new Error('Module does not exist');
+		}
+
 		const newTerm = new Term(term, definition, status, isStarred);
 
 		return await this.termRepository.createTerm(moduleId, newTerm);
 	}
 
 	async updateTerm(
+		userId: string,
 		moduleId: string,
 		termId: string,
 		{ term, definition, status, isStarred }: TermDto,
 	): Promise<TermModel | null> {
-		const existedModule = await this.termRepository.getTermById(moduleId, termId);
+		const existedModule = await this.moduleRepository.getModuleById(userId, moduleId);
 		if (!existedModule) {
-			return null;
+			throw new Error('Module does not exist');
 		}
+
+		const existedTerm = await this.termRepository.getTermById(moduleId, termId);
+		if (!existedTerm) {
+			throw new Error('Term to be updated does not exist');
+		}
+
 		return this.termRepository.updateTerm(moduleId, termId, {
 			term,
 			definition,
@@ -40,15 +55,35 @@ export class TermService implements ITermService {
 		});
 	}
 
-	async deleteTerm(moduleId: string, termId: string): Promise<void> {
+	async deleteTerm(userId: string, moduleId: string, termId: string): Promise<void> {
+		const existedModule = await this.moduleRepository.getModuleById(userId, moduleId);
+		if (!existedModule) {
+			throw new Error('Module does not exist');
+		}
+
+		const existedTerm = await this.termRepository.getTermById(moduleId, termId);
+		if (!existedTerm) {
+			throw new Error('Term to be deleted does not exist');
+		}
+
 		await this.termRepository.deleteTerm(moduleId, termId);
 	}
 
-	async getTerms(moduleId: string): Promise<TermModel[] | null> {
+	async getTerms(userId: string, moduleId: string): Promise<TermModel[] | null> {
+		const existedModule = await this.moduleRepository.getModuleById(userId, moduleId);
+		if (!existedModule) {
+			throw new Error('Module does not exist');
+		}
+
 		return await this.termRepository.getTerms(moduleId);
 	}
 
-	async getTermById(moduleId: string, termId: string): Promise<TermModel | null> {
+	async getTermById(userId: string, moduleId: string, termId: string): Promise<TermModel | null> {
+		const existedModule = await this.moduleRepository.getModuleById(userId, moduleId);
+		if (!existedModule) {
+			throw new Error('Module does not exist');
+		}
+
 		return await this.termRepository.getTermById(moduleId, termId);
 	}
 }
